backend/requirements.txt
asgiref==3.5.0
boto3==1.20.52
botocore==1.23.52
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==4.0.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
